requirements.txt
fluxer.py
ollama